ImageCaption/requirements.txt
streamlit~=1.22.0
requests~=2.21.0
pillow~=9.4.0
pygtrans~=1.5.2
tensorflow~=2.9.1
pandas~=1.3.5
numpy~=1.21.6
h5py~=2.9.0
simplejson~=3.19.1
opencv-python~=4.7.0.72